feat(admin): open dashboard/shows/calendar to hosts, merge Shows pages
- Move the Telegram-bot (Moafunkbot) notifications toggle off the Dashboard
  into a new admin-only /config page.
- Open Dashboard, Shows and Calendar to hosts (everyone except guests) and
  make the Dashboard the post-login landing page for them. Non-admins read
  the open /api/shows-overview instead of the admin-only /api/shows, so no
  backend role changes are needed.
- Merge the admin Shows page and the host 'My Shows' flow-select into one
  /shows page: a 'My Shows' section with streaming-prep actions plus an
  'All Shows' section with Upcoming/All/Past filters and create. Remove the
  separate 'My Show' nav item; /stream/select redirects to /shows.
- Guests land on the detail page of their assigned show (then into the
  stream flow); they cannot reach Dashboard/Shows/Calendar.
- Widen ShowList/MonthCalendar to a shared ScheduleItem type so they render
  data from either /api/shows or /api/shows-overview.
