+++ frontend/src/admin/components/ShowListItem.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import type { Show } from '../api';
+import type { ScheduleItem } from '../api';
 
 const props = defineProps<{
-  show: Show;
+  show: ScheduleItem;
   isPast: boolean;
 }>();
 
@@ -48,25 +48,32 @@ const daysUntil = getDaysUntil(props.show.date);
     <div class="list-show-date-col">
       <span class="list-show-date">{{ formatDateShort(show.date) }}</span>
       <span :class="['badge', 'days-badge', getDaysClass(daysUntil)]">
-        {{ daysUntil < 0 ? 'Past' : daysUntil + 'd' }} </span>
+        {{ daysUntil < 0 ? 'Past' : daysUntil + 'd' }}
+      </span>
     </div>
     <div class="list-show-info">
       <span class="list-show-title-row">
         <span class="list-show-title">{{ show.title }}</span>
-        <span v-if="show.show_type === 'unheard' || !show.show_type" :class="[
-          'badge',
-          'artist-badge',
-          {
-            'count-empty': show.artists.length === 0,
-            'count-partial': show.artists.length > 0 && show.artists.length < 4,
-            'count-full': show.artists.length >= 4,
-          },
-        ]">
+        <span
+          v-if="show.show_type === 'unheard' || !show.show_type"
+          :class="[
+            'badge',
+            'artist-badge',
+            {
+              'count-empty': show.artists.length === 0,
+              'count-partial': show.artists.length > 0 && show.artists.length < 4,
+              'count-full': show.artists.length >= 4,
+            },
+          ]"
+        >
           {{ show.artists.length }}/4
         </span>
       </span>
-      <span v-if="show.show_type === 'unheard' || !show.show_type" class="list-show-artists text-muted">
-        {{show.artists.map((a) => a.name).join(', ') || 'No artists assigned'}}
+      <span
+        v-if="show.show_type === 'unheard' || !show.show_type"
+        class="list-show-artists text-muted"
+      >
+        {{ show.artists.map((a) => a.name).join(', ') || 'No artists assigned' }}
       </span>
     </div>
     <div class="list-show-meta">
